[ssclient] Only exclude key file from VCS if it's under VCS control
Persuant to the previous commit, only add the key file generated/exported to the
VCS ignore file if it is in a directory currently under VCS control.

This only supports git for now, but is structured in a way that should make it
easy to add support for other VCS in the future.