Use separate function for retrieving arbitrary types
Rename the existing `SecretsManager::get()` to `SecretsManager::get_as()` and
introduce a new `SecretsManager::get()` that is a wrapper around
`SecretsManager::get_as::<String>()`.

This greatly simplifies the most common use case where the password is stored in
the vault as a Unicode string; no type hints are now needed.