[ssclient] Handle retrieval of binary secrets
While `ssclient` cannot save binary secrets, it can open stores that were
created/written to via the SecureStore API and contain binary secrets.
Previously, we would error out with a deserialization failure but now we catch
non-UTF-8 strings and instead return the base64 representation of the stored
(decrypted) binary bytes.

If retrieved as text, `base64:<encoded>` is returned. If retrieved as JSON, the
raw bytes are instead encoded (to avoid any ambiguities about the actual
contents).