Stop requiring store path in calls to `SecretsManager::new()`
This lets libraries abstract over `SecretsManager` and return an instance of
the type without specifying the path beforehand; the user can then call
`save_as()` to save the file to a path of their choosing afterwards.

This change is not expected to affect most users as the majority will be using
`ssclient` to create the store.