Add new workspace README.md
This is the "integrated" README, combining basic elements of both the `ssclient`
cli README and the `securestore` crate/API README in a holistic fashion.

This new README features a walkthrough of using both components to create a
secrets store, store some secrets, then retrieve them at runtime; as well as
some recommendations for best practices.