Add Send/Sync assertion for SecretsManager
We recommend using it as a global `static` via `Lazy<SecretsManager>` in the
repository README, so make sure it's actually possible to do that by asserting
that `SecretsManager` implements `Sync`.

We can probably live without the `Send` assertion, but if we break one we're
probably going to break the other (until we patch it with interior
synchronization and mutability).