Fix KeySource::File backwards compatibility
This was originally non-generic taking only `&Path` but then a release made it
generic over `P: AsRef<Path>`, breaking the literal usage of `KeySource::Csprng`
and `KeySource::Password` in the process (because the default generic fallback
didn't work, requiring `KeySource::<&Path>::Password` to be used instead; so we
reverted that and made it non-generic again.

To restore backwards compatibility, the non-generic `KeySource::File` has been
renamed to `KeySource::Path` and a new function masquerading as the old enum
variant `fn KeySource::File(path: AsRef<Path>)` has been added, and
`SecretsManager::load()` and `SecretsManager::new()` have been updated to take a
new trait `GenericKeySource` instead of a `KeySource` directly.

The backwards compatibility function-as-a-variant is hidden from the docs; a
publicly available and non-deprecated function `KeySource::from_file()` has also
been added that does the same thing and should be used going forward.

This patch also makes it possible to specify a `KeySource::Path` by passing a
string or a path directly to `new()` or `load()`. This wasn't an intentional
decision but rather naturally arose out of the approach used for backwards
compatibility and I think it's fine to keep.