Support loading a `SecretsManager` from arbitrary `Read` types
We can't have `SecretsManager::load()` directly take the `Read` parameter
because a) we are bound by backwards compatibility to support passing in a path
as the first parameter to `load()` and `Path` does not implement `Read`, b) we
would want to support passing in a path in all cases for ergonomics purposes.

Any arbitrary type may implement `GenericVaultSource` and provide a `Read` type
when `GenericVaultSource::read()` is called; the `SecretsManager` will be loaded
from this type rather from a path, allowing more abstract types of vaults not
backed by physical files to be used.

`GenericVaultSource` is not sealed and may be implemented by users of this
crate (it is now part of the public crate api/signature).

See discussion that incentivized this change in #3.

Note that it is still not possible to write out a `SecretsManager` to an
arbitrary type implementing `Write`; this can also be implemented but the
motivation for wanting to do so is currently unclear.