Add `KeySource::Buffer()`
In combination with previously making `GenericKeySource` public and unsealed,
this should be sufficient for downstream `SecureStore` users to load keys from
arbitrary sources.

It's difficult to add a `Read` variant of `KeySource` while keeping `KeySource`
functions immutable (taking `&self` rather than `&mut self` or `self`) as we
need to modify the `Read` object in order to get any data out of it.
Implementors will need to read the keys into a buffer to use `GenericKeySource`,
but that should be fine since keys are always limited in size.

Rust needs some sort of trait to represent "this immutable source can be read
from but doesn't implement `Read` itself", à la `IntoIterator`…